requirements.txt
Django<2.0
djangorestframework
requests
requests-cache
django-mptt
django-modeltranslation
six
pyyaml
unicodecsv
psycopg2
pytest-django==3.4.5
pytest==4.1.1             # via pytest-django
python-dateutil